move CodeID and ActorState to vm crate
