+++ vm/Cargo.toml
@@ -9,3 +9,5 @@ num-bigint = { path = "../math/bigint", package = "forest_bigint" }
 address = { package = "forest_address", path = "./address" }
 encoding = { package = "forest_encoding", path = "../encoding" }
 serde = { version = "1.0", features = ["derive"] }
+cid = { package = "forest_cid", path = "../ipld/cid" }
+
