+++ vm/actor/src/builtin/init.rs
@@ -1,8 +1,10 @@
 // Copyright 2020 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-use crate::{ActorID, CodeID};
-use vm::{ExitCode, MethodNum, Serialized, SysCode, METHOD_CONSTRUCTOR, METHOD_PLACEHOLDER};
+use vm::{
+    ActorID, CodeID, ExitCode, MethodNum, Serialized, SysCode, METHOD_CONSTRUCTOR,
+    METHOD_PLACEHOLDER,
+};
 
 use address::Address;
 use num_derive::FromPrimitive;
