+++ vm/src/lib.rs
@@ -1,11 +1,15 @@
 // Copyright 2020 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
+mod actor_state;
+mod code;
 mod exit_code;
 mod invoc;
 mod method;
 mod token;
 
+pub use self::actor_state::*;
+pub use self::code::*;
 pub use self::exit_code::*;
 pub use self::invoc::*;
 pub use self::method::*;
