+++ vm/state_tree/Cargo.toml
@@ -5,7 +5,6 @@ authors = ["ChainSafe Systems <info@chainsafe.io>"]
 edition = "2018"
 
 [dependencies]
-actor = { path = "../actor" }
 address = { package = "forest_address", path = "../address" }
 vm = { path = "../../vm" }
 cid = { package = "forest_cid", path = "../../ipld/cid" }
